pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.0",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
